refactor(ui): drop the account-section barrel
