--- packages/ui/src/mosaic/user-profile/user-profile-account-section/index.ts
@@ -1,17 +1,0 @@
-export { fill, userProfileAccountSectionBase } from './user-profile-account-section.messages';
-export type { UserProfileFormError } from './user-profile-account-section.types';
-export { UserProfileSaveError } from './user-profile-account-section.types';
-export type {
-  UserProfileAccountSectionViewProps,
-  UserProfileEmail,
-  UserProfilePhone,
-} from './user-profile-account-section.view';
-export { UserProfileAccountSectionView } from './user-profile-account-section.view';
-export type { UserProfileEditNameControllerOptions } from './user-profile-edit-name.controller';
-export { useUserProfileEditNameController } from './user-profile-edit-name.controller';
-export type {
-  UserProfileEditNameField,
-  UserProfileEditNameValue,
-  UserProfileEditNameViewProps,
-} from './user-profile-edit-name.view';
-export { UserProfileEditNameView } from './user-profile-edit-name.view';
